refactor(core): use SchemasService in HeaderClassExtractor
