+++ springwolf-core/src/test/java/io/github/springwolf/core/asyncapi/components/DefaultComponentsServiceTest.java
@@ -7,32 +7,18 @@ import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import io.github.springwolf.asyncapi.v3.model.channel.message.MessageObject;
-import io.github.springwolf.asyncapi.v3.model.schema.SchemaObject;
-import io.github.springwolf.core.asyncapi.components.postprocessors.SchemasPostProcessor;
+import io.github.springwolf.core.asyncapi.schemas.SchemaService;
 import io.github.springwolf.core.asyncapi.schemas.SwaggerSchemaUtil;
-import io.github.springwolf.core.configuration.properties.SpringwolfConfigProperties;
 import io.swagger.v3.core.util.Json;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 
 @ExtendWith(MockitoExtension.class)
 class DefaultComponentsServiceTest {
@@ -43,20 +29,13 @@ class DefaultComponentsServiceTest {
             new DefaultPrettyPrinter().withObjectIndenter(new DefaultIndenter("  ", DefaultIndenter.SYS_LF));
 
     @Mock
-    private SchemasPostProcessor schemasPostProcessor;
-
-    @Mock
-    private SchemasPostProcessor schemasPostProcessor2;
+    private SchemaService schemaService;
 
     private ComponentsService componentsService;
 
     @BeforeEach
     void setUp() {
-        componentsService = new DefaultComponentsService(
-                List.of(),
-                List.of(schemasPostProcessor, schemasPostProcessor2),
-                new SwaggerSchemaUtil(),
-                new SpringwolfConfigProperties());
+        componentsService = new DefaultComponentsService(new SwaggerSchemaUtil(), schemaService);
     }
 
     @Test
@@ -81,130 +60,5 @@ class DefaultComponentsServiceTest {
 
         System.out.println("Got: " + actualDefinitions);
         assertEquals(expected, actualDefinitions);
-    }
-
-    @Test
-    void classWithSchemaAnnotation() {
-        String modelName =
-                componentsService.registerSchema(ClassWithSchemaAnnotation.class, "content-type-not-relevant");
-
-        assertThat(modelName).isEqualTo("DifferentName");
-    }
-
-    @Test
-    void getDefinitionWithoutFqnClassName() throws IOException {
-        // given
-        SpringwolfConfigProperties properties = new SpringwolfConfigProperties();
-        properties.setUseFqn(false);
-
-        ComponentsService componentsServiceWithFqn =
-                new DefaultComponentsService(List.of(), List.of(), new SwaggerSchemaUtil(), properties);
-
-        // when
-        Class<?> clazz =
-                OneFieldFooWithoutFqn.class; // swagger seems to cache results. Therefore, a new class must be used.
-        componentsServiceWithFqn.registerSchema(clazz, "content-type-not-relevant");
-        String actualDefinitions =
-                objectMapper.writer(printer).writeValueAsString(componentsServiceWithFqn.getSchemas());
-
-        // then
-        System.out.println("Got: " + actualDefinitions);
-        assertThat(actualDefinitions).contains(clazz.getSimpleName());
-        assertThat(actualDefinitions).doesNotContain(clazz.getCanonicalName());
-    }
-
-    @Test
-    void postProcessorsAreCalled() {
-        componentsService.registerSchema(ClassWithSchemaAnnotation.class, "some-content-type");
-
-        verify(schemasPostProcessor).process(any(), any(), eq("some-content-type"));
-        verify(schemasPostProcessor2).process(any(), any(), eq("some-content-type"));
-    }
-
-    @Test
-    void postProcessorIsSkippedWhenSchemaWasRemoved() {
-        doAnswer(invocationOnMock -> {
-                    Map<String, Schema> schemas = invocationOnMock.getArgument(1);
-                    schemas.clear();
-                    return null;
-                })
-                .when(schemasPostProcessor)
-                .process(any(), any(), any());
-
-        componentsService.registerSchema(ClassWithSchemaAnnotation.class, "content-type-not-relevant");
-
-        verifyNoInteractions(schemasPostProcessor2);
-    }
-
-    @Nested
-    class RegisterSchema {
-
-        @Test
-        void Integer() {
-            // when
-            componentsService.registerSchema(Integer.class, "content-type-not-relevant");
-
-            // then
-            Map<String, SchemaObject> schemas = componentsService.getSchemas();
-            assertThat(schemas).hasSize(1).containsKey("java.lang.Number");
-            assertThat(schemas.get("java.lang.Number").getType()).isEqualTo("number");
-        }
-
-        @Test
-        void Double() {
-            // when
-            componentsService.registerSchema(Double.class, "content-type-not-relevant");
-
-            // then
-            Map<String, SchemaObject> schemas = componentsService.getSchemas();
-            assertThat(schemas).hasSize(1).containsKey("java.lang.Number");
-            assertThat(schemas.get("java.lang.Number").getType()).isEqualTo("number");
-        }
-
-        @Test
-        void String() {
-            // when
-            componentsService.registerSchema(String.class, "content-type-not-relevant");
-
-            // then
-            Map<String, SchemaObject> schemas = componentsService.getSchemas();
-            assertThat(schemas).hasSize(1).containsKey("java.lang.String");
-            assertThat(schemas.get("java.lang.String").getType()).isEqualTo("string");
-        }
-
-        @Test
-        void Byte() {
-            // when
-            componentsService.registerSchema(Byte.class, "content-type-not-relevant");
-
-            // then
-            Map<String, SchemaObject> schemas = componentsService.getSchemas();
-            assertThat(schemas).hasSize(1).containsKey("java.lang.String");
-            assertThat(schemas.get("java.lang.String").getType()).isEqualTo("string");
-        }
-
-        @Test
-        void Boolean() {
-            // when
-            componentsService.registerSchema(Boolean.class, "content-type-not-relevant");
-
-            // then
-            Map<String, SchemaObject> schemas = componentsService.getSchemas();
-            assertThat(schemas).hasSize(1).containsKey("java.lang.Boolean");
-        }
-    }
-
-    @Data
-    @NoArgsConstructor
-    @Schema(name = "DifferentName")
-    private static class ClassWithSchemaAnnotation {
-        private String s;
-        private boolean b;
-    }
-
-    @Data
-    @NoArgsConstructor
-    private static class OneFieldFooWithoutFqn {
-        private String s;
     }
 }
