+++ springwolf-core/src/test/java/io/github/springwolf/core/asyncapi/schemas/SchemaObjectMergerTest.java
@@ -4,13 +4,16 @@ package io.github.springwolf.core.asyncapi.schemas;
 import io.github.springwolf.asyncapi.v3.model.schema.SchemaObject;
 import org.junit.jupiter.api.Test;
 
+import java.util.HashMap;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class SchemaObjectMergerTest {
 
     @Test
     void merge() {
-        SchemaObject initial = new SchemaObject();
+        SchemaObject initial =
+                SchemaObject.builder().properties(new HashMap<>()).build();
         initial.getProperties().put("key1", "value1");
 
         // when
@@ -22,7 +25,8 @@ class SchemaObjectMergerTest {
 
     @Test
     void mergeAndIgnoreNullValues() {
-        SchemaObject initial = new SchemaObject();
+        SchemaObject initial =
+                SchemaObject.builder().properties(new HashMap<>()).build();
         initial.getProperties().put("key1", "value1");
 
         // when
@@ -34,15 +38,18 @@ class SchemaObjectMergerTest {
 
     @Test
     void mergeWhileNotOverwritingInitialValues() {
-        SchemaObject initial = new SchemaObject();
+        SchemaObject initial =
+                SchemaObject.builder().properties(new HashMap<>()).build();
         initial.getProperties().put("key1", "value1");
 
-        SchemaObject schema1 = new SchemaObject();
-        initial.getProperties().put("key1", "value2");
+        SchemaObject schema1 =
+                SchemaObject.builder().properties(new HashMap<>()).build();
+        schema1.getProperties().put("key1", "value2");
         schema1.getProperties().put("key2", "value2");
 
-        SchemaObject schema2 = new SchemaObject();
-        schema1.getProperties().put("key2", "value3");
+        SchemaObject schema2 =
+                SchemaObject.builder().properties(new HashMap<>()).build();
+        schema2.getProperties().put("key2", "value3");
         schema2.getProperties().put("key3", "value3");
 
         // when
